rename main fld to game
